Working test setup still w/o CRs, but w/o IPSec loops.
In previous setup the subnet ranges of the RAs were also advertised via BGP.
That causes problems to a peer when sending already encapsulated packets to
the other peer, since its dest IP is part of what should be encapsulated.
Now RAs get their IPs from a specific subnet not included in advertisements.
Also for convenience it is separate from assumed GCP block to allow for
summarization.