Non completely working setup with CRs talking BGP to RAs.
The CRs talk BGP to RAs. They announce the hub subnets for now, no custom
advertisement set. However after the announcements some routes are deleted
in the RAs (e.g. the default route) and things stop working.
This is one of the multiple tests done, all with the same results regarding
routes deleted in VyOS. In this case, there are no more "special" subnets
in the hub nor onprem for RAs and CRs. Previous IPSec loops with this setup
are avoided setting static routes to peers in the RAs.